Add a better documentation structure
Also added folders for markdown and pdf, so that I can easily change
between different documents. ALso wanted to be able to eventually do the
thing Karl was doing with gimli and the makefile, where I can specify
special types of .css files if I wanted, so it's nice to have more
folders to do things like that.