Adding more support for the server
This contains the user keys that I need for the server. It also contains
a script to easily switch between encoded and decoded options.

I have attached the current users.enc, which contains the information
necessary for the keys.

Make sure to store user files in a *.no format, so that git does not
include them.

Also added a quick command for starting the task server. May be expanded
at a later point to do more for me