Created a functional prototype of pdf splitting
This will take a pdf, and then split it page by page. After splitting it
into the different slides, it will insert the references of those slides
into a markdown document so that when I take notes I have a pre-existing
book of slides that I can take my notes under.

Also added some pip support and a test script for the document.