Remove auto update prompt
I might put this back later, but for now the prompt looks fine as it is.
I don't really know if I need the time ticking up every second... Could
maybe set the timeout for one minute or something like that and have it
just show the minute, not the time.